+++ src/Listeners/LogSqlQuery.php
@@ -4,9 +4,10 @@ namespace Onlime\LaravelSqlReporter\Listeners;
 
 use Illuminate\Console\Events\CommandFinished;
 use Illuminate\Foundation\Http\Events\RequestHandled;
+use Illuminate\Support\Facades\App;
 use Onlime\LaravelSqlReporter\SqlLogger;
 
-class LogSqlQuery
+readonly class LogSqlQuery
 {
     /**
      * Create the event listener.
@@ -21,6 +22,12 @@ class LogSqlQuery
      */
     public function handle(CommandFinished|RequestHandled $event): void
     {
+        // Prevent duplicate logging when running in request (RequestHandled event)
+        // and programmatically executing Artisan commands (CommandFinished event).
+        if ($event instanceof RequestHandled === App::runningInConsole()) {
+            return;
+        }
+
         $this->logger->log();
     }
 }